requirements.txt
contourpy==1.3.1
cycler==0.12.1
darkdetect==0.7.1
fonttools==4.55.2
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
numpy-stl==3.2.0
packaging==24.2
pillow==11.0.0
pygame==2.6.1